Revises Project Structure And Documentation
This commit revises the project structure by spliting models, services and utilities into seperate sub folders. This commit also adds development documentation to the project.
