+++ src/models/connection.model.js
@@ -66,6 +66,7 @@ class Connection extends EmbeddedDocument {
       }
     });
   }
+
   /**
    * preInit is a hook
    * @schema
@@ -73,12 +74,14 @@ class Connection extends EmbeddedDocument {
    * @param {Object} data The data used to create the client.
    * @return {null} The preInit hook does not return anything.
    */
+
   preInit(data) {
     this.credentials = Credentials.create({
       user: data.user,
       password: data.password
     });
   }
+
   /**
    * @method _basicAuth
    * @private
@@ -88,6 +91,7 @@ class Connection extends EmbeddedDocument {
    * @return {String} A string containing the user and password authentication
    * pair.
    */
+
   _basicAuth() {
     const auth = `Basic ${new Buffer(
       `${this.credentials.user}:${this.credentials.password}`
@@ -104,6 +108,7 @@ class Connection extends EmbeddedDocument {
    * @return {String} a token retrieved from the private generation method
    *
    */
+
   _saveToken(data) {
     this.expires = moment()
       .add(15, 'minutes')
@@ -112,6 +117,7 @@ class Connection extends EmbeddedDocument {
     this.token = data.response.token;
     return data;
   }
+
   /**
    * @method valid
    * @public
@@ -121,12 +127,14 @@ class Connection extends EmbeddedDocument {
    * @return {String} a token retrieved from the private generation method
    *
    */
+
   valid() {
     return (
       this.token !== undefined &&
       moment().isBetween(this.issued, this.expires, '()')
     );
   }
+
   /**
    * @method generate
    * @memberof Connection
@@ -137,6 +145,7 @@ class Connection extends EmbeddedDocument {
    * @return {Promise} returns a promise that will either resolve or reject based on the Data API.
    * response.
    */
+
   generate(axios, url) {
     return new Promise((resolve, reject) =>
       axios
@@ -155,6 +164,7 @@ class Connection extends EmbeddedDocument {
         .catch(error => reject(error))
     );
   }
+
   /**
    * @method clears
    * @memberof Connection
@@ -165,6 +175,7 @@ class Connection extends EmbeddedDocument {
    * @return {Object} response The response recieved from the Data API.
    *
    */
+
   clear(response) {
     this.token = '';
     this.issued = '';
@@ -182,6 +193,7 @@ class Connection extends EmbeddedDocument {
    * @return {Object} response The response recieved from the Data API.
    *
    */
+
   extend(response) {
     this.expires = moment()
       .add(15, 'minutes')
