+++ src/models/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const { Client } = require('./client.model');
+
+module.exports = { Client };
