+++ src/services/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const { fieldData, recordId, transform } = require('./transform.service');
+
+module.exports = { fieldData, recordId, transform };
