+++ src/utilities/conversion.utilities.js
@@ -85,6 +85,7 @@ const omit = (data, properties) =>
  * @param  {Any} values The value to attempt to parse.
  * @return {Object|Any} A json object or array of objects without the properties passed to it
  */
+
 const parse = value => (isJson(value) ? JSON.parse(value) : value);
 
 module.exports = {
